# eyeguard/OpenAPI.yaml
...
openapi: 3.0.3
info:
  title: EyeGuard API
  version: 1.0.0
  description: |
    Software-only simulation REST API for the EyeGuard SOC training platform.
servers:
  - url: https://localhost:8000
    description: Local FastAPI dev server
paths:
...
  /api/v1/settings/users:
    get:
      summary: List active users
      operationId: listUsers
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Submit signup request
      operationId: requestSignup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignupRequest'
      responses:
        '202':
          description: Signup queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSignupQueued'
...
  /api/v1/settings/users/{requestId}/approve:
    post:
      summary: Approve signup
      operationId: approveSignup
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/v1/settings/users/{requestId}/reject:
    post:
      summary: Reject signup
      operationId: rejectSignup
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Signup rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRejection'
components:
  parameters:
...
    UserId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
    UserSignupRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
        role:
          type: string
    UserSignupQueued:
      type: object
      properties:
        request_id:
          type: string
        status:
          type: string
    UserRejection:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        message:
          type: string
...